revert https://github.com/commonwarexyz/monorepo/pull/3778
