+++ examples/sync/src/bin/client.rs
@@ -11,11 +11,7 @@ use commonware_runtime::{
 };
 use commonware_storage::{
     mmr,
-    qmdb::{
-        any::sync::Target,
-        current as current_qmdb,
-        sync::{self, compact},
-    },
+    qmdb::sync::{self, compact},
 };
 use commonware_sync::{
     any, crate_version, current,
@@ -74,9 +70,9 @@ struct Config {
 async fn target_update_task<E, Op, D>(
     context: E,
     resolver: Resolver<Op, D>,
-    update_tx: mpsc::Sender<Target<mmr::Family, D>>,
+    update_tx: mpsc::Sender<sync::Target<mmr::Family, D>>,
     interval_duration: Duration,
-    initial_target: Target<mmr::Family, D>,
+    initial_target: sync::Target<mmr::Family, D>,
 ) -> Result<(), Error>
 where
     E: Clock,
@@ -136,8 +132,8 @@ where
         E,
         Config,
         Resolver<Op, Key>,
-        Target<mmr::Family, Key>,
-        mpsc::Receiver<Target<mmr::Family, Key>>,
+        sync::Target<mmr::Family, Key>,
+        mpsc::Receiver<sync::Target<mmr::Family, Key>>,
         u32,
     ) -> SyncFut,
     SyncFut: Future<Output = Result<DB, Box<dyn std::error::Error>>>,
@@ -192,7 +188,7 @@ where
         |context, config, resolver, initial_target, update_receiver, iteration| async move {
             let db_config = any::create_config(&context);
             let sync_config =
-                sync::engine::Config::<any::Database<_>, Resolver<any::Operation, Key>, _> {
+                sync::engine::Config::<any::Database<_>, Resolver<any::Operation, Key>> {
                     context,
                     db_config,
                     fetch_batch_size: config.batch_size,
@@ -221,94 +217,44 @@ where
 
 /// Repeatedly sync a Current database to the server's state.
 ///
-/// Uses the `current::sync::sync` wrapper. The wrapper verifies each target's `OpsRootWitness`
-/// before forwarding its ops root to the shared sync engine, then checks the database root for the
-/// target the engine finishes on.
+/// The sync engine targets the ops root (not the canonical root). After sync completes,
+/// the bitmap and grafted MMR are reconstructed from the synced operations.
 async fn run_current<E>(context: E, config: Config) -> Result<(), Box<dyn std::error::Error>>
 where
     E: BufferPooler + Storage + Clock + Metrics + Network + Spawner,
 {
-    info!("starting Current database sync process");
-    let mut iteration = 0u32;
-    loop {
-        let resolver =
-            Resolver::<current::Operation, Key>::connect(context.child("resolver"), config.server)
-                .await?;
-
-        let initial_target = resolver.get_current_sync_target().await?;
-        info!(
-            root = %initial_target.root,
-            ops_root = %initial_target.ops_root,
-            range = ?initial_target.range,
-            "received current sync target"
-        );
-
-        let (update_sender, update_receiver) = mpsc::channel(UPDATE_CHANNEL_SIZE);
-
-        let target_update_handle = {
-            let resolver = resolver.clone();
-            let mut current_target_root = initial_target.root;
-            let target_update_interval = config.target_update_interval;
-            context
-                .child("target_update")
-                .spawn(move |context| async move {
-                    loop {
-                        context.sleep(target_update_interval).await;
-                        match resolver.get_current_sync_target().await {
-                            Ok(new_target) => {
-                                if current_target_root != new_target.root {
-                                    let new_root = new_target.root;
-                                    match update_sender.clone().try_send(new_target) {
-                                        Ok(()) => {
-                                            info!("target updated");
-                                            current_target_root = new_root;
-                                        }
-                                        Err(mpsc::error::TrySendError::Closed(_)) => return Ok(()),
-                                        Err(err) => {
-                                            warn!(?err, "failed to send target update");
-                                            return Err(Error::TargetUpdateChannel {
-                                                reason: err.to_string(),
-                                            });
-                                        }
-                                    }
-                                }
-                            }
-                            Err(err) => {
-                                warn!(?err, "failed to get sync target from server");
-                            }
-                        }
-                    }
-                })
-        };
-
-        let db_config = current::create_config(&context);
-        let database: current::Database<_> = current_qmdb::sync::sync(current_qmdb::sync::Config {
-            context: context.child("sync"),
-            resolver,
-            target: initial_target,
-            max_outstanding_requests: config.max_outstanding_requests,
-            fetch_batch_size: config.batch_size,
-            apply_batch_size: 1024,
-            db_config,
-            update_rx: Some(update_receiver),
-            finish_rx: None,
-            reached_target_tx: None,
-            max_retained_roots: 8,
-        })
-        .await?;
-
-        target_update_handle.abort();
-        info!(
-            sync_iteration = iteration,
-            root = %database.root(),
-            ops_root = %database.ops_root(),
-            sync_interval = ?config.sync_interval,
-            "Current sync completed successfully"
-        );
-        database.destroy().await?;
-        context.sleep(config.sync_interval).await;
-        iteration += 1;
-    }
+    run_full_sync::<current::Database<_>, current::Operation, _, _, _>(
+        context,
+        config,
+        |context, config, resolver, initial_target, update_receiver, iteration| async move {
+            let db_config = current::create_config(&context);
+            let sync_config =
+                sync::engine::Config::<current::Database<_>, Resolver<current::Operation, Key>> {
+                    context,
+                    db_config,
+                    fetch_batch_size: config.batch_size,
+                    target: initial_target,
+                    resolver,
+                    apply_batch_size: 1024,
+                    max_outstanding_requests: config.max_outstanding_requests,
+                    update_rx: Some(update_receiver),
+                    finish_rx: None,
+                    reached_target_tx: None,
+                    max_retained_roots: 8,
+                };
+            let database: current::Database<_> = sync::sync(sync_config).await?;
+            info!(
+                sync_iteration = iteration,
+                canonical_root = %database.root(),
+                ops_root = %database.ops_root(),
+                sync_interval = ?config.sync_interval,
+                "Current sync completed successfully"
+            );
+            Ok(database)
+        },
+        "Current database",
+    )
+    .await
 }
 
 /// Repeatedly sync an Immutable database to the server's state.
@@ -324,7 +270,6 @@ where
             let sync_config = sync::engine::Config::<
                 immutable::Database<_>,
                 Resolver<immutable::Operation, Key>,
-                _,
             > {
                 context,
                 db_config,
@@ -362,23 +307,20 @@ where
         config,
         |context, config, resolver, initial_target, update_receiver, iteration| async move {
             let db_config = keyless::create_config(&context);
-            let sync_config = sync::engine::Config::<
-                keyless::Database<_>,
-                Resolver<keyless::Operation, Key>,
-                _,
-            > {
-                context,
-                db_config,
-                fetch_batch_size: config.batch_size,
-                target: initial_target,
-                resolver,
-                apply_batch_size: 1024,
-                max_outstanding_requests: config.max_outstanding_requests,
-                update_rx: Some(update_receiver),
-                finish_rx: None,
-                reached_target_tx: None,
-                max_retained_roots: 8,
-            };
+            let sync_config =
+                sync::engine::Config::<keyless::Database<_>, Resolver<keyless::Operation, Key>> {
+                    context,
+                    db_config,
+                    fetch_batch_size: config.batch_size,
+                    target: initial_target,
+                    resolver,
+                    apply_batch_size: 1024,
+                    max_outstanding_requests: config.max_outstanding_requests,
+                    update_rx: Some(update_receiver),
+                    finish_rx: None,
+                    reached_target_tx: None,
+                    max_retained_roots: 8,
+                };
             let database: keyless::Database<_> = sync::sync(sync_config).await?;
             info!(
                 sync_iteration = iteration,
