+++ examples/sync/src/net/wire.rs
@@ -7,11 +7,7 @@ use commonware_runtime::{Buf, BufMut};
 use commonware_storage::{
     merkle::MAX_PINNED_NODES,
     mmr::{self, Location, Proof},
-    qmdb::{
-        any::sync::Target,
-        current::sync::Target as CurrentTarget,
-        sync::{compact, compact::State},
-    },
+    qmdb::sync::{compact, compact::State, Target},
 };
 use std::num::NonZeroU64;
 
@@ -54,16 +50,6 @@ where
 {
     pub request_id: RequestId,
     pub target: Target<mmr::Family, D>,
-}
-
-/// Response with current-database sync target (database root + witness).
-#[derive(Debug)]
-pub struct GetCurrentSyncTargetResponse<D>
-where
-    D: Digest,
-{
-    pub request_id: RequestId,
-    pub target: CurrentTarget<mmr::Family, D>,
 }
 
 /// Request for compact authenticated state.
@@ -116,7 +102,6 @@ where
     GetCompactTargetResponse(GetCompactTargetResponse<D>),
     GetCompactStateRequest(GetCompactStateRequest<D>),
     GetCompactStateResponse(GetCompactStateResponse<Op, D>),
-    GetCurrentSyncTargetResponse(GetCurrentSyncTargetResponse<D>),
     Error(ErrorResponse),
 }
 
@@ -134,7 +119,6 @@ where
             Self::GetCompactTargetResponse(r) => r.request_id,
             Self::GetCompactStateRequest(r) => r.request_id,
             Self::GetCompactStateResponse(r) => r.request_id,
-            Self::GetCurrentSyncTargetResponse(r) => r.request_id,
             Self::Error(e) => e.request_id,
         }
     }
@@ -190,10 +174,6 @@ where
                 7u8.write(buf);
                 resp.write(buf);
             }
-            Self::GetCurrentSyncTargetResponse(resp) => {
-                9u8.write(buf);
-                resp.write(buf);
-            }
             Self::Error(err) => {
                 8u8.write(buf);
                 err.write(buf);
@@ -217,7 +197,6 @@ where
             Self::GetCompactTargetResponse(resp) => resp.encode_size(),
             Self::GetCompactStateRequest(req) => req.encode_size(),
             Self::GetCompactStateResponse(resp) => resp.encode_size(),
-            Self::GetCurrentSyncTargetResponse(resp) => resp.encode_size(),
             Self::Error(err) => err.encode_size(),
         }
     }
@@ -254,9 +233,6 @@ where
                 GetCompactStateResponse::read(buf)?,
             )),
             8 => Ok(Self::Error(ErrorResponse::read(buf)?)),
-            9 => Ok(Self::GetCurrentSyncTargetResponse(
-                GetCurrentSyncTargetResponse::read(buf)?,
-            )),
             d => Err(CodecError::InvalidEnum(d)),
         }
     }
@@ -433,28 +409,6 @@ where
     fn read_cfg(buf: &mut impl Buf, _: &()) -> Result<Self, CodecError> {
         let request_id = RequestId::read_cfg(buf, &())?;
         let target = Target::<mmr::Family, D>::read_cfg(buf, &())?;
-        Ok(Self { request_id, target })
-    }
-}
-
-impl<D: Digest> Write for GetCurrentSyncTargetResponse<D> {
-    fn write(&self, buf: &mut impl BufMut) {
-        self.request_id.write(buf);
-        self.target.write(buf);
-    }
-}
-
-impl<D: Digest> EncodeSize for GetCurrentSyncTargetResponse<D> {
-    fn encode_size(&self) -> usize {
-        self.request_id.encode_size() + self.target.encode_size()
-    }
-}
-
-impl<D: Digest> Read for GetCurrentSyncTargetResponse<D> {
-    type Cfg = ();
-    fn read_cfg(buf: &mut impl Buf, _: &()) -> Result<Self, CodecError> {
-        let request_id = RequestId::read_cfg(buf, &())?;
-        let target = CurrentTarget::<mmr::Family, D>::read_cfg(buf, &())?;
         Ok(Self { request_id, target })
     }
 }
