+++ storage/src/qmdb/current/sync/tests.rs
@@ -1,21 +1,22 @@
 //! Tests for [crate::qmdb::current] state sync.
 //!
-//! This module reuses the shared sync test functions from [crate::qmdb::sync::tests] by
-//! implementing [SyncTestHarness] for current database types.
+//! This module reuses the shared sync test functions from [crate::qmdb::any::sync::tests] by
+//! implementing [SyncTestHarness] for current database types. The key difference from `any`
+//! harnesses is that `sync_target_root` returns the **QMDB ops root** (via
+//! [qmdb::sync::Database::root](crate::qmdb::sync::Database::root)), not the canonical root
+//! returned by `Db::root()`.
 //!
 //! Harnesses are instantiated for **both** MMR and MMB merkle families across each (ordered,
 //! unordered) x (fixed, variable) database variant, so the shared suite runs twice per variant.
 //!
 //! In addition to the shared harness-based suite, this module contains focused tests for
-//! `current`-specific sync behavior: overlay-state authentication (database-root check), pruned
+//! `current`-specific sync behavior: overlay-state authentication (canonical-root check), pruned
 //! MMB round-trip, and target-update regression coverage.
 
 use crate::qmdb::{
+    any::sync::tests::{ConfigOf, SyncTestHarness},
     current::tests::{fixed_config, variable_config},
-    sync::{
-        tests::{ConfigOf, SyncTestHarness},
-        Database as SyncDatabase,
-    },
+    sync::Database as SyncDatabase,
 };
 use commonware_cryptography::{sha256::Digest, Sha256};
 use commonware_macros::test_traced;
@@ -23,34 +24,14 @@ use commonware_parallel::Sequential;
 use commonware_runtime::{
     deterministic, deterministic::Context, BufferPooler, Runner as _, Supervisor as _,
 };
-use commonware_utils::{non_empty_range, range::NonEmptyRange};
+use commonware_utils::non_empty_range;
 use rand::RngCore as _;
-
-fn dummy_current_target<F: crate::merkle::Graftable>(
-    root: Digest,
-    ops_root: Digest,
-    range: NonEmptyRange<crate::merkle::Location<F>>,
-) -> crate::qmdb::current::sync::Target<F, Digest> {
-    crate::qmdb::current::sync::Target::new(
-        root,
-        ops_root,
-        crate::qmdb::current::proof::OpsRootWitness {
-            grafted_root: Digest::from([0; 32]),
-            pending_chunk_digest: None.try_into().unwrap(),
-            partial_chunk: None,
-        },
-        range,
-    )
-}
 
 // ===== Harness Implementations =====
 
 mod harnesses {
     use super::*;
-    use crate::{
-        merkle::{self, mmb, mmr},
-        qmdb::current::{proof::OpsRootWitness, sync::Target as CurrentTarget},
-    };
+    use crate::merkle::{self, mmb, mmr};
     use commonware_math::algebra::Random;
     use commonware_utils::test_rng_seeded;
 
@@ -94,23 +75,6 @@ mod harnesses {
         32,
         Sequential,
     >;
-
-    fn target<F: merkle::Graftable>(
-        root: Digest,
-        ops_root: Digest,
-        range: NonEmptyRange<crate::merkle::Location<F>>,
-    ) -> CurrentTarget<F, Digest> {
-        CurrentTarget::new(
-            root,
-            ops_root,
-            OpsRootWitness {
-                grafted_root: Digest::from([0; 32]),
-                pending_chunk_digest: None.try_into().unwrap(),
-                partial_chunk: None,
-            },
-            range,
-        )
-    }
 
     fn create_unordered_fixed_ops<F: merkle::Family>(
         n: usize,
@@ -312,14 +276,9 @@ mod harnesses {
     impl<F: merkle::Graftable> SyncTestHarness for UnorderedFixedHarness<F> {
         type Family = F;
         type Db = UnorderedFixedDb<F>;
-        type Target = CurrentTarget<F, Digest>;
 
-        fn target(
-            root: Digest,
-            ops_root: Digest,
-            range: NonEmptyRange<crate::merkle::Location<Self::Family>>,
-        ) -> Self::Target {
-            target(root, ops_root, range)
+        fn sync_target_root(db: &Self::Db) -> Digest {
+            SyncDatabase::root(db)
         }
 
         fn config(suffix: &str, pooler: &impl BufferPooler) -> ConfigOf<Self> {
@@ -364,14 +323,9 @@ mod harnesses {
     impl<F: merkle::Graftable> SyncTestHarness for UnorderedVariableHarness<F> {
         type Family = F;
         type Db = UnorderedVariableDb<F>;
-        type Target = CurrentTarget<F, Digest>;
 
-        fn target(
-            root: Digest,
-            ops_root: Digest,
-            range: NonEmptyRange<crate::merkle::Location<Self::Family>>,
-        ) -> Self::Target {
-            target(root, ops_root, range)
+        fn sync_target_root(db: &Self::Db) -> Digest {
+            SyncDatabase::root(db)
         }
 
         fn config(suffix: &str, pooler: &impl BufferPooler) -> ConfigOf<Self> {
@@ -416,14 +370,9 @@ mod harnesses {
     impl<F: merkle::Graftable> SyncTestHarness for OrderedFixedHarness<F> {
         type Family = F;
         type Db = OrderedFixedDb<F>;
-        type Target = CurrentTarget<F, Digest>;
 
-        fn target(
-            root: Digest,
-            ops_root: Digest,
-            range: NonEmptyRange<crate::merkle::Location<Self::Family>>,
-        ) -> Self::Target {
-            target(root, ops_root, range)
+        fn sync_target_root(db: &Self::Db) -> Digest {
+            SyncDatabase::root(db)
         }
 
         fn config(suffix: &str, pooler: &impl BufferPooler) -> ConfigOf<Self> {
@@ -468,14 +417,9 @@ mod harnesses {
     impl<F: merkle::Graftable> SyncTestHarness for OrderedVariableHarness<F> {
         type Family = F;
         type Db = OrderedVariableDb<F>;
-        type Target = CurrentTarget<F, Digest>;
 
-        fn target(
-            root: Digest,
-            ops_root: Digest,
-            range: NonEmptyRange<crate::merkle::Location<Self::Family>>,
-        ) -> Self::Target {
-            target(root, ops_root, range)
+        fn sync_target_root(db: &Self::Db) -> Digest {
+            SyncDatabase::root(db)
         }
 
         fn config(suffix: &str, pooler: &impl BufferPooler) -> ConfigOf<Self> {
@@ -517,12 +461,12 @@ mod harnesses {
 }
 
 /// Regression test: sync a pruned MMB-backed current DB and verify the synced DB has the
-/// same database root, reopens cleanly, and returns the expected value.
+/// same canonical root, reopens cleanly, and returns the expected value.
 ///
 /// The target DB commits the same key 100 times, forcing the inactivity floor past a full
 /// 256-bit chunk boundary. Without overlay-state in the sync protocol, the receiver
 /// re-derives `pruned_chunks` from `range.start / chunk_bits` and builds a grafted tree
-/// whose pinned nodes don't match the sender's. The database roots diverge.
+/// whose pinned nodes don't match the sender's. The canonical roots diverge.
 #[test_traced("INFO")]
 fn test_current_mmb_sync_with_pruned_full_chunk_reopens() {
     let executor = deterministic::Runner::default();
@@ -570,7 +514,7 @@ fn test_current_mmb_sync_with_pruned_full_chunk_reopens() {
 
         target_db.prune(target_db.sync_boundary()).await.unwrap();
 
-        let sync_root = SyncDatabase::ops_root(&target_db);
+        let sync_root = SyncDatabase::root(&target_db);
         let verification_root = target_db.root();
         let lower_bound = target_db.sync_boundary();
         let upper_bound = target_db.bounds().await.end;
@@ -578,17 +522,17 @@ fn test_current_mmb_sync_with_pruned_full_chunk_reopens() {
         let client_suffix = context.next_u64().to_string();
         let client_config = variable_config::<crate::translator::TwoCap>(&client_suffix, &context);
         let target_db = std::sync::Arc::new(target_db);
-        // This uses the shared sync engine directly. The focused `root_sync` tests below cover the
-        // current sync wrapper that authenticates target witnesses before starting the engine.
+        // Supply the trusted canonical root so `build_db`'s authentication check actually
+        // runs: this is the success-path coverage for the overlay-state authentication
+        // anchor. A bad-root rejection path test belongs with the focused sync tests.
         let synced_db: Db = crate::qmdb::sync::sync(crate::qmdb::sync::engine::Config {
             context: context.child("client"),
             db_config: client_config.clone(),
             fetch_batch_size: commonware_utils::NZU64!(64),
-            target: dummy_current_target(
-                verification_root,
-                sync_root,
-                commonware_utils::non_empty_range!(lower_bound, upper_bound),
-            ),
+            target: crate::qmdb::sync::Target {
+                root: sync_root,
+                range: commonware_utils::non_empty_range!(lower_bound, upper_bound),
+            },
             resolver: target_db.clone(),
             apply_batch_size: 1024,
             max_outstanding_requests: 4,
@@ -600,7 +544,7 @@ fn test_current_mmb_sync_with_pruned_full_chunk_reopens() {
         .await
         .unwrap();
 
-        assert_eq!(SyncDatabase::ops_root(&synced_db), sync_root);
+        assert_eq!(SyncDatabase::root(&synced_db), sync_root);
         assert_eq!(synced_db.root(), verification_root);
         assert_eq!(synced_db.sync_boundary(), lower_bound);
         assert_eq!(synced_db.get(&key).await.unwrap(), expected);
@@ -610,7 +554,7 @@ fn test_current_mmb_sync_with_pruned_full_chunk_reopens() {
         let reopened: Db = Db::init(context.child("reopened"), client_config)
             .await
             .unwrap();
-        assert_eq!(SyncDatabase::ops_root(&reopened), sync_root);
+        assert_eq!(SyncDatabase::root(&reopened), sync_root);
         assert_eq!(reopened.root(), verification_root);
         assert_eq!(reopened.sync_boundary(), lower_bound);
         assert_eq!(reopened.get(&key).await.unwrap(), expected);
@@ -661,16 +605,14 @@ fn test_current_has_local_target_state_rejects_target_before_local_lower_bound()
         let bounds = db.bounds().await;
         let local_start = bounds.start;
         let local_end = bounds.end;
-        let sync_root = SyncDatabase::ops_root(&db);
+        let sync_root = SyncDatabase::root(&db);
 
         assert!(local_start > crate::merkle::Location::new(0));
 
-        let root = db.root();
-        let stale_target = dummy_current_target(
-            root,
-            sync_root,
-            non_empty_range!(local_start.checked_sub(1).unwrap(), local_end),
-        );
+        let stale_target = crate::qmdb::sync::Target {
+            root: sync_root,
+            range: non_empty_range!(local_start.checked_sub(1).unwrap(), local_end),
+        };
         assert!(
             !<Db as SyncDatabase>::has_local_target_state(
                 context.child("probe_stale"),
@@ -680,8 +622,10 @@ fn test_current_has_local_target_state_rejects_target_before_local_lower_bound()
             .await
         );
 
-        let matching_target =
-            dummy_current_target(root, sync_root, non_empty_range!(local_start, local_end));
+        let matching_target = crate::qmdb::sync::Target {
+            root: sync_root,
+            range: non_empty_range!(local_start, local_end),
+        };
         assert!(
             <Db as SyncDatabase>::has_local_target_state(
                 context.child("probe_matching"),
@@ -697,7 +641,7 @@ fn test_current_has_local_target_state_rejects_target_before_local_lower_bound()
 
 // ===== Test Generation Macro =====
 
-/// Dispatches to the shared test functions in [crate::qmdb::sync::tests].
+/// Dispatches to the shared test functions in [crate::qmdb::any::sync::tests].
 macro_rules! current_sync_tests_for_harness {
     ($harness:ty, $mod_name:ident) => {
         mod $mod_name {
@@ -708,7 +652,7 @@ macro_rules! current_sync_tests_for_harness {
 
             #[test_traced]
             fn test_sync_resolver_fails() {
-                crate::qmdb::sync::tests::test_sync_resolver_fails::<$harness>();
+                crate::qmdb::any::sync::tests::test_sync_resolver_fails::<$harness>();
             }
 
             #[rstest]
@@ -721,7 +665,7 @@ macro_rules! current_sync_tests_for_harness {
             #[case::db_size_eq_batch_size(1000, 1000)]
             #[case::batch_size_gt_db_size(1000, 1001)]
             fn test_sync(#[case] target_db_ops: usize, #[case] fetch_batch_size: u64) {
-                crate::qmdb::sync::tests::test_sync::<$harness>(
+                crate::qmdb::any::sync::tests::test_sync::<$harness>(
                     target_db_ops,
                     NonZeroU64::new(fetch_batch_size).unwrap(),
                 );
@@ -729,57 +673,67 @@ macro_rules! current_sync_tests_for_harness {
 
             #[test_traced]
             fn test_sync_subset_of_target_database() {
-                crate::qmdb::sync::tests::test_sync_subset_of_target_database::<$harness>(1000);
+                crate::qmdb::any::sync::tests::test_sync_subset_of_target_database::<$harness>(
+                    1000,
+                );
             }
 
             #[test_traced]
             fn test_sync_use_existing_db_partial_match() {
-                crate::qmdb::sync::tests::test_sync_use_existing_db_partial_match::<$harness>(1000);
+                crate::qmdb::any::sync::tests::test_sync_use_existing_db_partial_match::<$harness>(
+                    1000,
+                );
             }
 
             #[test_traced]
             fn test_sync_use_existing_db_exact_match() {
-                crate::qmdb::sync::tests::test_sync_use_existing_db_exact_match::<$harness>(1000);
+                crate::qmdb::any::sync::tests::test_sync_use_existing_db_exact_match::<$harness>(
+                    1000,
+                );
             }
 
             #[test_traced("WARN")]
             fn test_target_update_lower_bound_decrease() {
-                crate::qmdb::sync::tests::test_target_update_lower_bound_decrease::<$harness>();
+                crate::qmdb::any::sync::tests::test_target_update_lower_bound_decrease::<$harness>(
+                );
             }
 
             #[test_traced("WARN")]
             fn test_target_update_upper_bound_decrease() {
-                crate::qmdb::sync::tests::test_target_update_upper_bound_decrease::<$harness>();
+                crate::qmdb::any::sync::tests::test_target_update_upper_bound_decrease::<$harness>(
+                );
             }
 
             #[test_traced("WARN")]
             fn test_target_update_bounds_increase() {
-                crate::qmdb::sync::tests::test_target_update_bounds_increase::<$harness>();
+                crate::qmdb::any::sync::tests::test_target_update_bounds_increase::<$harness>();
             }
 
             #[test_traced("WARN")]
             fn test_target_update_on_done_client() {
-                crate::qmdb::sync::tests::test_target_update_on_done_client::<$harness>();
+                crate::qmdb::any::sync::tests::test_target_update_on_done_client::<$harness>();
             }
 
             #[test_traced]
             fn test_sync_waits_for_explicit_finish() {
-                crate::qmdb::sync::tests::test_sync_waits_for_explicit_finish::<$harness>();
+                crate::qmdb::any::sync::tests::test_sync_waits_for_explicit_finish::<$harness>();
             }
 
             #[test_traced]
             fn test_sync_handles_early_finish_signal() {
-                crate::qmdb::sync::tests::test_sync_handles_early_finish_signal::<$harness>();
+                crate::qmdb::any::sync::tests::test_sync_handles_early_finish_signal::<$harness>();
             }
 
             #[test_traced]
             fn test_sync_fails_when_finish_sender_dropped() {
-                crate::qmdb::sync::tests::test_sync_fails_when_finish_sender_dropped::<$harness>();
+                crate::qmdb::any::sync::tests::test_sync_fails_when_finish_sender_dropped::<
+                    $harness,
+                >();
             }
 
             #[test_traced]
             fn test_sync_allows_dropped_reached_target_receiver() {
-                crate::qmdb::sync::tests::test_sync_allows_dropped_reached_target_receiver::<
+                crate::qmdb::any::sync::tests::test_sync_allows_dropped_reached_target_receiver::<
                     $harness,
                 >();
             }
@@ -801,7 +755,7 @@ macro_rules! current_sync_tests_for_harness {
                 #[case] initial_ops: usize,
                 #[case] additional_ops: usize,
             ) {
-                crate::qmdb::sync::tests::test_target_update_during_sync::<$harness>(
+                crate::qmdb::any::sync::tests::test_target_update_during_sync::<$harness>(
                     initial_ops,
                     additional_ops,
                 );
@@ -809,12 +763,12 @@ macro_rules! current_sync_tests_for_harness {
 
             #[test_traced]
             fn test_sync_database_persistence() {
-                crate::qmdb::sync::tests::test_sync_database_persistence::<$harness>();
+                crate::qmdb::any::sync::tests::test_sync_database_persistence::<$harness>();
             }
 
             #[test_traced]
             fn test_sync_post_sync_usability() {
-                crate::qmdb::sync::tests::test_sync_post_sync_usability::<$harness>();
+                crate::qmdb::any::sync::tests::test_sync_post_sync_usability::<$harness>();
             }
         }
     };
@@ -834,309 +788,3 @@ current_sync_tests_for_harness!(harnesses::OrderedFixedMmrHarness, ordered_fixed
 current_sync_tests_for_harness!(harnesses::OrderedFixedMmbHarness, ordered_fixed_mmb);
 current_sync_tests_for_harness!(harnesses::OrderedVariableMmrHarness, ordered_variable_mmr);
 current_sync_tests_for_harness!(harnesses::OrderedVariableMmbHarness, ordered_variable_mmb);
-
-mod root_sync {
-    use super::*;
-    use crate::{
-        merkle::mmr,
-        qmdb::{
-            self,
-            current::{
-                proof::OpsRootWitness,
-                sync::{self as current_sync, Target as CurrentTarget},
-                tests::variable_config,
-            },
-        },
-    };
-    use commonware_runtime::{Clock, Spawner};
-    use commonware_utils::NZU64;
-
-    type Db = crate::qmdb::current::unordered::variable::Db<
-        mmr::Family,
-        Context,
-        Digest,
-        Digest,
-        Sha256,
-        crate::translator::TwoCap,
-        32,
-        Sequential,
-    >;
-
-    async fn apply_round(db: &mut Db, key: Digest, round: u64) {
-        let merkleized = db
-            .new_batch()
-            .write(key, Some(Digest::from([round as u8; 32])))
-            .merkleize(db, None)
-            .await
-            .unwrap();
-        db.apply_batch(merkleized).await.unwrap();
-        db.commit().await.unwrap();
-    }
-
-    async fn build_target_db(context: &mut Context) -> Db {
-        let suffix = context.next_u64().to_string();
-        let cfg = variable_config::<crate::translator::TwoCap>(&suffix, context);
-        let mut db: Db = Db::init(context.child("target"), cfg).await.unwrap();
-
-        let key = Digest::from([7u8; 32]);
-        for round in 0..10u64 {
-            apply_round(&mut db, key, round).await;
-        }
-        db.sync().await.unwrap();
-        db
-    }
-
-    async fn make_current_target(db: &Db) -> CurrentTarget<mmr::Family, Digest> {
-        let hasher = qmdb::hasher::<Sha256>();
-        let witness = db.ops_root_witness(&hasher).await.unwrap();
-        let lower = db.sync_boundary();
-        let upper = db.bounds().await.end;
-        CurrentTarget::new(
-            db.root(),
-            db.ops_root(),
-            witness,
-            non_empty_range!(lower, upper),
-        )
-    }
-
-    #[test_traced("INFO")]
-    fn test_root_sync_succeeds() {
-        let executor = deterministic::Runner::default();
-        executor.start(|mut context: Context| async move {
-            let target_db = build_target_db(&mut context).await;
-            let target = make_current_target(&target_db).await;
-            let root = target.root;
-
-            let client_suffix = context.next_u64().to_string();
-            let client_config =
-                variable_config::<crate::translator::TwoCap>(&client_suffix, &context);
-
-            let target_db = std::sync::Arc::new(target_db);
-            let synced_db: Db = current_sync::sync(current_sync::Config {
-                context: context.child("client"),
-                resolver: target_db.clone(),
-                target,
-                max_outstanding_requests: 4,
-                fetch_batch_size: NZU64!(64),
-                apply_batch_size: 1024,
-                db_config: client_config,
-                update_rx: None,
-                finish_rx: None,
-                reached_target_tx: None,
-                max_retained_roots: 8,
-            })
-            .await
-            .unwrap();
-
-            assert_eq!(synced_db.root(), root);
-
-            synced_db.destroy().await.unwrap();
-            let target_db = std::sync::Arc::into_inner(target_db).unwrap();
-            target_db.destroy().await.unwrap();
-        });
-    }
-
-    #[test_traced("INFO")]
-    fn test_root_sync_tracks_target_update() {
-        let executor = deterministic::Runner::default();
-        executor.start(|mut context: Context| async move {
-            let mut target_db = build_target_db(&mut context).await;
-            let initial_target = make_current_target(&target_db).await;
-
-            let key = Digest::from([7u8; 32]);
-            for round in 10..20u64 {
-                apply_round(&mut target_db, key, round).await;
-            }
-            target_db.sync().await.unwrap();
-            let updated_target = make_current_target(&target_db).await;
-            let expected_root = updated_target.root;
-
-            let (update_sender, update_receiver) = commonware_utils::channel::mpsc::channel(1);
-            let (finish_sender, finish_receiver) = commonware_utils::channel::mpsc::channel(1);
-            update_sender.send(updated_target).await.unwrap();
-            drop(update_sender);
-            context.child("finish").spawn(move |context| async move {
-                context.sleep(std::time::Duration::from_millis(1)).await;
-                finish_sender.send(()).await.unwrap();
-            });
-
-            let client_suffix = context.next_u64().to_string();
-            let client_config =
-                variable_config::<crate::translator::TwoCap>(&client_suffix, &context);
-            let target_db = std::sync::Arc::new(target_db);
-
-            let synced_db: Db = current_sync::sync(current_sync::Config {
-                context: context.child("client"),
-                resolver: target_db.clone(),
-                target: initial_target,
-                max_outstanding_requests: 1,
-                fetch_batch_size: NZU64!(1),
-                apply_batch_size: 1024,
-                db_config: client_config,
-                update_rx: Some(update_receiver),
-                finish_rx: Some(finish_receiver),
-                reached_target_tx: None,
-                max_retained_roots: 8,
-            })
-            .await
-            .unwrap();
-
-            assert_eq!(synced_db.root(), expected_root);
-
-            synced_db.destroy().await.unwrap();
-            let target_db = std::sync::Arc::into_inner(target_db).unwrap();
-            target_db.destroy().await.unwrap();
-        });
-    }
-
-    #[test_traced("INFO")]
-    fn test_root_sync_rejects_invalid_witness() {
-        let executor = deterministic::Runner::default();
-        executor.start(|mut context: Context| async move {
-            let target_db = build_target_db(&mut context).await;
-            let mut target = make_current_target(&target_db).await;
-
-            target.witness = OpsRootWitness {
-                grafted_root: Digest::from([0xFFu8; 32]),
-                ..target.witness
-            };
-
-            let client_suffix = context.next_u64().to_string();
-            let client_config =
-                variable_config::<crate::translator::TwoCap>(&client_suffix, &context);
-
-            let target_db = std::sync::Arc::new(target_db);
-            let result: Result<Db, _> = current_sync::sync(current_sync::Config {
-                context: context.child("client"),
-                resolver: target_db.clone(),
-                target,
-                max_outstanding_requests: 4,
-                fetch_batch_size: NZU64!(64),
-                apply_batch_size: 1024,
-                db_config: client_config,
-                update_rx: None,
-                finish_rx: None,
-                reached_target_tx: None,
-                max_retained_roots: 8,
-            })
-            .await;
-
-            assert!(matches!(
-                result,
-                Err(crate::qmdb::sync::Error::Engine(
-                    crate::qmdb::sync::EngineError::OpsRootWitnessInvalid
-                ))
-            ));
-
-            let target_db = std::sync::Arc::into_inner(target_db).unwrap();
-            target_db.destroy().await.unwrap();
-        });
-    }
-
-    #[test_traced("INFO")]
-    fn test_root_sync_rejects_invalid_update_witness() {
-        let executor = deterministic::Runner::default();
-        executor.start(|mut context: Context| async move {
-            let mut target_db = build_target_db(&mut context).await;
-            let initial_target = make_current_target(&target_db).await;
-
-            let key = Digest::from([7u8; 32]);
-            for round in 10..20u64 {
-                apply_round(&mut target_db, key, round).await;
-            }
-            target_db.sync().await.unwrap();
-            let mut updated_target = make_current_target(&target_db).await;
-            updated_target.witness = OpsRootWitness {
-                grafted_root: Digest::from([0xFFu8; 32]),
-                ..updated_target.witness
-            };
-
-            let (update_sender, update_receiver) = commonware_utils::channel::mpsc::channel(1);
-            let (_finish_sender, finish_receiver) = commonware_utils::channel::mpsc::channel(1);
-            update_sender.send(updated_target).await.unwrap();
-            drop(update_sender);
-
-            let client_suffix = context.next_u64().to_string();
-            let client_config =
-                variable_config::<crate::translator::TwoCap>(&client_suffix, &context);
-            let target_db = std::sync::Arc::new(target_db);
-
-            let result: Result<Db, _> = current_sync::sync(current_sync::Config {
-                context: context.child("client"),
-                resolver: target_db.clone(),
-                target: initial_target,
-                max_outstanding_requests: 1,
-                fetch_batch_size: NZU64!(1),
-                apply_batch_size: 1024,
-                db_config: client_config,
-                update_rx: Some(update_receiver),
-                finish_rx: Some(finish_receiver),
-                reached_target_tx: None,
-                max_retained_roots: 8,
-            })
-            .await;
-
-            assert!(matches!(
-                result,
-                Err(crate::qmdb::sync::Error::Engine(
-                    crate::qmdb::sync::EngineError::OpsRootWitnessInvalid
-                ))
-            ));
-
-            let target_db = std::sync::Arc::into_inner(target_db).unwrap();
-            target_db.destroy().await.unwrap();
-        });
-    }
-
-    /// Verify that closing the caller's update channel while the engine is waiting
-    /// for updates (with finish_rx) allows the engine to eventually complete.
-    /// This exercises the Either::Right path in the forwarding select: the forward
-    /// future exits, update_tx must be dropped so the engine sees
-    /// UpdateChannelClosed, and then finish completes sync.
-    #[test_traced("INFO")]
-    fn test_root_sync_update_channel_close_unblocks_engine() {
-        let executor = deterministic::Runner::default();
-        executor.start(|mut context: Context| async move {
-            let target_db = build_target_db(&mut context).await;
-            let target = make_current_target(&target_db).await;
-            let root = target.root;
-
-            let client_suffix = context.next_u64().to_string();
-            let client_config =
-                variable_config::<crate::translator::TwoCap>(&client_suffix, &context);
-
-            let target_db = std::sync::Arc::new(target_db);
-
-            let (update_tx, update_rx) = commonware_utils::channel::mpsc::channel(1);
-            let (finish_tx, finish_rx) = commonware_utils::channel::mpsc::channel(1);
-
-            let sync_fut = current_sync::sync(current_sync::Config {
-                context: context.child("client"),
-                resolver: target_db.clone(),
-                target,
-                max_outstanding_requests: 4,
-                fetch_batch_size: NZU64!(64),
-                apply_batch_size: 1024,
-                db_config: client_config,
-                update_rx: Some(update_rx),
-                finish_rx: Some(finish_rx),
-                reached_target_tx: None,
-                max_retained_roots: 8,
-            });
-
-            // Drop the update sender to close the channel while the engine is
-            // waiting for updates or a finish signal.
-            drop(update_tx);
-
-            // Send the finish signal so the engine can complete.
-            finish_tx.send(()).await.unwrap();
-
-            let synced_db: Db = sync_fut.await.unwrap();
-            assert_eq!(synced_db.root(), root);
-
-            synced_db.destroy().await.unwrap();
-            let target_db = std::sync::Arc::into_inner(target_db).unwrap();
-            target_db.destroy().await.unwrap();
-        });
-    }
-}
