+++ storage/src/qmdb/immutable/sync/tests.rs
@@ -10,13 +10,12 @@ use crate::{
     merkle::{self, full::Config as MerkleConfig, Location},
     qmdb::{
         self,
-        any::sync::Target,
         immutable::{self, variable::Operation},
         sync::{
             self,
             engine::{Config, NextStep},
             resolver::Resolver,
-            Engine,
+            Engine, Target,
         },
     },
     translator::TwoCap,
@@ -127,10 +126,10 @@ where
         let config = Config {
             db_config: db_config.clone(),
             fetch_batch_size,
-            target: Target::new(
-                target_root,
-                non_empty_range!(target_oldest_retained_loc, target_op_count),
-            ),
+            target: Target {
+                root: target_root,
+                range: non_empty_range!(target_oldest_retained_loc, target_op_count),
+            },
             context: context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -189,10 +188,10 @@ where
         let config = Config {
             db_config,
             fetch_batch_size: NZU64!(10),
-            target: Target::new(
-                target_root,
-                non_empty_range!(target_oldest_retained_loc, target_op_count),
-            ),
+            target: Target {
+                root: target_root,
+                range: non_empty_range!(target_oldest_retained_loc, target_op_count),
+            },
             context: context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -240,7 +239,10 @@ where
         let config = Config {
             db_config: db_config.clone(),
             fetch_batch_size: NZU64!(5),
-            target: Target::new(target_root, non_empty_range!(lower_bound, op_count)),
+            target: Target {
+                root: target_root,
+                range: non_empty_range!(lower_bound, op_count),
+            },
             context: client_context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -310,10 +312,10 @@ where
             let config = Config {
                 context: context.child("client"),
                 db_config: H::config(&format!("update_test_{}", context.next_u64()), &context),
-                target: Target::new(
-                    initial_root,
-                    non_empty_range!(initial_lower_bound, initial_upper_bound),
-                ),
+                target: Target {
+                    root: initial_root,
+                    range: non_empty_range!(initial_lower_bound, initial_upper_bound),
+                },
                 resolver: target_db.clone(),
                 fetch_batch_size: NZU64!(2),
                 max_outstanding_requests: 10,
@@ -323,12 +325,11 @@ where
                 reached_target_tx: None,
                 max_retained_roots: 1,
             };
-            let mut client: Engine<DbOf<H>, _, Target<H::Family, sha256::Digest>> =
-                Engine::new(config).await.unwrap();
+            let mut client: Engine<DbOf<H>, _> = Engine::new(config).await.unwrap();
             loop {
                 client = match client.step().await.unwrap() {
                     NextStep::Continue(new_client) => new_client,
-                    NextStep::Complete(..) => panic!("client should not be complete"),
+                    NextStep::Complete(_) => panic!("client should not be complete"),
                 };
                 let log_size = Contiguous::size(client.journal()).await;
                 if log_size > *initial_lower_bound {
@@ -338,10 +339,10 @@ where
         };
 
         update_sender
-            .send(Target::new(
-                final_root,
-                non_empty_range!(initial_lower_bound, final_upper_bound),
-            ))
+            .send(Target {
+                root: final_root,
+                range: non_empty_range!(initial_lower_bound, final_upper_bound),
+            })
             .await
             .unwrap();
 
@@ -393,7 +394,10 @@ where
         let config = Config {
             db_config: H::config(&format!("subset_{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(10),
-            target: Target::new(target_root, non_empty_range!(lower_bound, op_count)),
+            target: Target {
+                root: target_root,
+                range: non_empty_range!(lower_bound, op_count),
+            },
             context: context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -445,7 +449,10 @@ where
         let config = Config {
             db_config: sync_db_config,
             fetch_batch_size: NZU64!(10),
-            target: Target::new(root, non_empty_range!(lower_bound, upper_bound)),
+            target: Target {
+                root,
+                range: non_empty_range!(lower_bound, upper_bound),
+            },
             context: context.child("sync"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -495,7 +502,10 @@ where
         let config = Config {
             db_config: sync_config,
             fetch_batch_size: NZU64!(10),
-            target: Target::new(root, non_empty_range!(lower_bound, upper_bound)),
+            target: Target {
+                root,
+                range: non_empty_range!(lower_bound, upper_bound),
+            },
             context: context.child("sync"),
             resolver: resolver.clone(),
             apply_batch_size: 1024,
@@ -541,10 +551,10 @@ where
             context: context.child("client"),
             db_config: H::config(&format!("lb-dec-{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(5),
-            target: Target::new(
-                initial_root,
-                non_empty_range!(initial_lower_bound, initial_upper_bound),
-            ),
+            target: Target {
+                root: initial_root,
+                range: non_empty_range!(initial_lower_bound, initial_upper_bound),
+            },
             resolver: target_db.clone(),
             apply_batch_size: 1024,
             max_outstanding_requests: 10,
@@ -553,17 +563,16 @@ where
             reached_target_tx: None,
             max_retained_roots: 1,
         };
-        let client: Engine<DbOf<H>, _, Target<H::Family, sha256::Digest>> =
-            Engine::new(config).await.unwrap();
+        let client: Engine<DbOf<H>, _> = Engine::new(config).await.unwrap();
 
         update_sender
-            .send(Target::new(
-                initial_root,
-                non_empty_range!(
+            .send(Target {
+                root: initial_root,
+                range: non_empty_range!(
                     initial_lower_bound.checked_sub(1).unwrap(),
                     initial_upper_bound
                 ),
-            ))
+            })
             .await
             .unwrap();
 
@@ -603,10 +612,10 @@ where
             context: context.child("client"),
             db_config: H::config(&format!("ub-dec-{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(5),
-            target: Target::new(
-                initial_root,
-                non_empty_range!(initial_lower_bound, initial_upper_bound),
-            ),
+            target: Target {
+                root: initial_root,
+                range: non_empty_range!(initial_lower_bound, initial_upper_bound),
+            },
             resolver: target_db.clone(),
             apply_batch_size: 1024,
             max_outstanding_requests: 10,
@@ -615,14 +624,13 @@ where
             reached_target_tx: None,
             max_retained_roots: 1,
         };
-        let client: Engine<DbOf<H>, _, Target<H::Family, sha256::Digest>> =
-            Engine::new(config).await.unwrap();
+        let client: Engine<DbOf<H>, _> = Engine::new(config).await.unwrap();
 
         update_sender
-            .send(Target::new(
-                initial_root,
-                non_empty_range!(initial_lower_bound, initial_upper_bound - 1),
-            ))
+            .send(Target {
+                root: initial_root,
+                range: non_empty_range!(initial_lower_bound, initial_upper_bound - 1),
+            })
             .await
             .unwrap();
 
@@ -676,10 +684,10 @@ where
             context: context.child("client"),
             db_config: H::config(&format!("bounds_inc_{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(1),
-            target: Target::new(
-                initial_root,
-                non_empty_range!(initial_lower_bound, initial_upper_bound),
-            ),
+            target: Target {
+                root: initial_root,
+                range: non_empty_range!(initial_lower_bound, initial_upper_bound),
+            },
             resolver: target_db.clone(),
             apply_batch_size: 1024,
             max_outstanding_requests: 1,
@@ -690,10 +698,10 @@ where
         };
 
         update_sender
-            .send(Target::new(
-                final_root,
-                non_empty_range!(final_lower_bound, final_upper_bound),
-            ))
+            .send(Target {
+                root: final_root,
+                range: non_empty_range!(final_lower_bound, final_upper_bound),
+            })
             .await
             .unwrap();
 
@@ -747,7 +755,10 @@ where
         let config = Config {
             db_config,
             fetch_batch_size: NZU64!(100),
-            target: Target::new(target_root, non_empty_range!(bounds.start, bounds.end)),
+            target: Target {
+                root: target_root,
+                range: non_empty_range!(bounds.start, bounds.end),
+            },
             context: context.child("client"),
             resolver: target_db.clone(),
             apply_batch_size: 1024,
@@ -808,7 +819,10 @@ where
             context: context.child("client"),
             db_config: H::config(&format!("done_{}", context.next_u64()), &context),
             fetch_batch_size: NZU64!(20),
-            target: Target::new(root, non_empty_range!(lower_bound, upper_bound)),
+            target: Target {
+                root,
+                range: non_empty_range!(lower_bound, upper_bound),
+            },
             resolver: target_db.clone(),
             apply_batch_size: 1024,
             max_outstanding_requests: 10,
@@ -821,10 +835,10 @@ where
         let synced_db: DbOf<H> = sync::sync(config).await.unwrap();
 
         let _ = update_sender
-            .send(Target::new(
-                sha256::Digest::from([2u8; 32]),
-                non_empty_range!(lower_bound + 1, upper_bound + 1),
-            ))
+            .send(Target {
+                root: sha256::Digest::from([2u8; 32]),
+                range: non_empty_range!(lower_bound + 1, upper_bound + 1),
+            })
             .await;
 
         assert_eq!(H::db_root(&synced_db), root);
@@ -1222,8 +1236,10 @@ mod compact_variable_mmr {
         deterministic::Runner::default().start(|_context| async move {
             let resolver: Arc<commonware_utils::sync::AsyncRwLock<Option<SourceDb>>> =
                 Arc::new(commonware_utils::sync::AsyncRwLock::new(None));
-            let target =
-                sync::compact::Target::new(sha256::Digest::from([0; 32]), Location::new(1));
+            let target = sync::compact::Target {
+                root: sha256::Digest::from([0; 32]),
+                leaf_count: Location::new(1),
+            };
 
             assert!(matches!(
                 sync::compact::Resolver::get_compact_state(&resolver, target).await,
@@ -1253,7 +1269,10 @@ mod compact_variable_mmr {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target::new(source.root(), bounds.end);
+            let target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: bounds.end,
+            };
             let source = Arc::new(source);
             let client_cfg = client_config(&suffix);
             let client: ClientDb = sync::compact::sync(sync::compact::Config {
@@ -1299,7 +1318,10 @@ mod compact_variable_mmr {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target::new(source.root(), bounds.end);
+            let target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: bounds.end,
+            };
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1343,7 +1365,10 @@ mod compact_variable_mmr {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target::new(source.root(), bounds.end);
+            let target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: bounds.end,
+            };
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1398,7 +1423,10 @@ mod compact_variable_mmr {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target::new(source.root(), bounds.end);
+            let target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: bounds.end,
+            };
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1440,7 +1468,10 @@ mod compact_variable_mmr {
                 .merkleize(&source, Some(vec![1]), Location::new(1));
             source.apply_batch(batch1).await.unwrap();
             source.commit().await.unwrap();
-            let stale_target = sync::compact::Target::new(source.root(), source.bounds().await.end);
+            let stale_target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: source.bounds().await.end,
+            };
 
             let batch2 = source
                 .new_batch()
@@ -1448,8 +1479,10 @@ mod compact_variable_mmr {
                 .merkleize(&source, Some(vec![2]), Location::new(2));
             source.apply_batch(batch2).await.unwrap();
             source.commit().await.unwrap();
-            let current_target =
-                sync::compact::Target::new(source.root(), source.bounds().await.end);
+            let current_target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: source.bounds().await.end,
+            };
             assert_ne!(stale_target, current_target);
 
             let source = Arc::new(source);
@@ -1682,8 +1715,10 @@ mod compact_variable_mmb {
         deterministic::Runner::default().start(|_context| async move {
             let resolver: Arc<commonware_utils::sync::AsyncRwLock<Option<SourceDb>>> =
                 Arc::new(commonware_utils::sync::AsyncRwLock::new(None));
-            let target =
-                sync::compact::Target::new(sha256::Digest::from([0; 32]), Location::new(1));
+            let target = sync::compact::Target {
+                root: sha256::Digest::from([0; 32]),
+                leaf_count: Location::new(1),
+            };
 
             assert!(matches!(
                 sync::compact::Resolver::get_compact_state(&resolver, target).await,
@@ -1713,7 +1748,10 @@ mod compact_variable_mmb {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target::new(source.root(), bounds.end);
+            let target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: bounds.end,
+            };
             let source = Arc::new(source);
             let client_cfg = client_config(&suffix);
             let client: ClientDb = sync::compact::sync(sync::compact::Config {
@@ -1759,7 +1797,10 @@ mod compact_variable_mmb {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target::new(source.root(), bounds.end);
+            let target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: bounds.end,
+            };
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1803,7 +1844,10 @@ mod compact_variable_mmb {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target::new(source.root(), bounds.end);
+            let target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: bounds.end,
+            };
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1858,7 +1902,10 @@ mod compact_variable_mmb {
             source.commit().await.unwrap();
 
             let bounds = source.bounds().await;
-            let target = sync::compact::Target::new(source.root(), bounds.end);
+            let target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: bounds.end,
+            };
             let source = Arc::new(source);
             let mut state = sync::compact::Resolver::get_compact_state(&source, target.clone())
                 .await
@@ -1900,7 +1947,10 @@ mod compact_variable_mmb {
                 .merkleize(&source, Some(vec![1]), Location::new(1));
             source.apply_batch(batch1).await.unwrap();
             source.commit().await.unwrap();
-            let stale_target = sync::compact::Target::new(source.root(), source.bounds().await.end);
+            let stale_target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: source.bounds().await.end,
+            };
 
             let batch2 = source
                 .new_batch()
@@ -1908,8 +1958,10 @@ mod compact_variable_mmb {
                 .merkleize(&source, Some(vec![2]), Location::new(2));
             source.apply_batch(batch2).await.unwrap();
             source.commit().await.unwrap();
-            let current_target =
-                sync::compact::Target::new(source.root(), source.bounds().await.end);
+            let current_target = sync::compact::Target {
+                root: source.root(),
+                leaf_count: source.bounds().await.end,
+            };
             assert_ne!(stale_target, current_target);
 
             let source = Arc::new(source);
