+++ storage/src/qmdb/keyless/sync/mod.rs
@@ -121,10 +121,6 @@ where
         Ok(db)
     }
 
-    fn ops_root(&self) -> Self::Digest {
-        self.root()
-    }
-
     fn root(&self) -> Self::Digest {
         self.root()
     }
