+++ storage/src/qmdb/sync/compact.rs
@@ -92,11 +92,6 @@ pub struct Target<F: Family, D: Digest> {
 impl<F: Family, D: Digest> Target<F, D> {
     const INVALID_LEAF_COUNT: &'static str = "leaf_count must be in 1..=MAX_LEAVES";
 
-    /// Create a compact-sync target.
-    pub const fn new(root: D, leaf_count: Location<F>) -> Self {
-        Self { root, leaf_count }
-    }
-
     /// Validate a compact target that may have been constructed programmatically.
     pub fn validate(&self) -> Result<(), &'static str> {
         if !self.leaf_count.is_valid() || self.leaf_count == 0 {
@@ -517,7 +512,12 @@ macro_rules! impl_compact_resolver_keyless {
             ) -> Result<FetchResult<Self::Family, Self::Op, Self::Digest>, Self::Error> {
                 fetch_state_from_full_source(
                     target,
-                    || async { Target::new(self.root(), self.bounds().await.end) },
+                    || async {
+                        Target {
+                            root: self.root(),
+                            leaf_count: self.bounds().await.end,
+                        }
+                    },
                     |leaf_count, last_commit_loc| {
                         self.historical_proof(
                             leaf_count,
@@ -552,7 +552,12 @@ macro_rules! impl_compact_resolver_keyless {
                 let db = self.read().await;
                 fetch_state_from_full_source(
                     target,
-                    || async { Target::new(db.root(), db.bounds().await.end) },
+                    || async {
+                        Target {
+                            root: db.root(),
+                            leaf_count: db.bounds().await.end,
+                        }
+                    },
                     |leaf_count, last_commit_loc| {
                         db.historical_proof(
                             leaf_count,
@@ -588,7 +593,12 @@ macro_rules! impl_compact_resolver_keyless {
                 let db = guard.as_ref().ok_or(ServeError::MissingSource)?;
                 fetch_state_from_full_source(
                     target,
-                    || async { Target::new(db.root(), db.bounds().await.end) },
+                    || async {
+                        Target {
+                            root: db.root(),
+                            leaf_count: db.bounds().await.end,
+                        }
+                    },
                     |leaf_count, last_commit_loc| {
                         db.historical_proof(
                             leaf_count,
@@ -631,7 +641,12 @@ macro_rules! impl_compact_resolver_immutable {
             ) -> Result<FetchResult<Self::Family, Self::Op, Self::Digest>, Self::Error> {
                 fetch_state_from_full_source(
                     target,
-                    || async { Target::new(self.root(), self.bounds().await.end) },
+                    || async {
+                        Target {
+                            root: self.root(),
+                            leaf_count: self.bounds().await.end,
+                        }
+                    },
                     |leaf_count, last_commit_loc| {
                         self.historical_proof(
                             leaf_count,
@@ -669,7 +684,12 @@ macro_rules! impl_compact_resolver_immutable {
                 let db = self.read().await;
                 fetch_state_from_full_source(
                     target,
-                    || async { Target::new(db.root(), db.bounds().await.end) },
+                    || async {
+                        Target {
+                            root: db.root(),
+                            leaf_count: db.bounds().await.end,
+                        }
+                    },
                     |leaf_count, last_commit_loc| {
                         db.historical_proof(
                             leaf_count,
@@ -708,7 +728,12 @@ macro_rules! impl_compact_resolver_immutable {
                 let db = guard.as_ref().ok_or(ServeError::MissingSource)?;
                 fetch_state_from_full_source(
                     target,
-                    || async { Target::new(db.root(), db.bounds().await.end) },
+                    || async {
+                        Target {
+                            root: db.root(),
+                            leaf_count: db.bounds().await.end,
+                        }
+                    },
                     |leaf_count, last_commit_loc| {
                         db.historical_proof(
                             leaf_count,
@@ -1072,33 +1097,5 @@ mod tests {
             }
             assert_eq!(attempts.load(Ordering::SeqCst), 1);
         });
-    }
-}
-
-#[cfg(feature = "arbitrary")]
-impl<F: Family, D: Digest> arbitrary::Arbitrary<'_> for Target<F, D>
-where
-    D: for<'a> arbitrary::Arbitrary<'a>,
-{
-    fn arbitrary(u: &mut arbitrary::Unstructured<'_>) -> arbitrary::Result<Self> {
-        let root = u.arbitrary()?;
-        let leaf_count = u.int_in_range(1..=*F::MAX_LEAVES)?;
-        Ok(Self {
-            root,
-            leaf_count: Location::new(leaf_count),
-        })
-    }
-}
-
-#[cfg(all(test, feature = "arbitrary"))]
-mod conformance {
-    use super::*;
-    use crate::merkle::{mmb, mmr};
-    use commonware_codec::conformance::CodecConformance;
-    use commonware_cryptography::sha256::Digest as Sha256Digest;
-
-    commonware_conformance::conformance_tests! {
-        CodecConformance<Target<mmr::Family, Sha256Digest>>,
-        CodecConformance<Target<mmb::Family, Sha256Digest>>,
     }
 }
