+++ storage/src/qmdb/sync/database.rs
@@ -1,6 +1,6 @@
 use crate::{
     merkle::{Family, Location},
-    qmdb::sync::{Journal, Target},
+    qmdb::sync::Journal,
     translator::Translator,
 };
 use commonware_cryptography::Digest;
@@ -64,21 +64,15 @@ pub trait Database: Sized + Send {
     /// on-disk database already reflects the target. Simple append-only variants may
     /// verify only the persisted tree size and root. Variants with additional
     /// pruning-dependent state should also ensure their persisted lower bound still
-    /// covers `target.range().start()`.
-    fn has_local_target_state<T>(
+    /// covers `target.range.start()`.
+    fn has_local_target_state(
         _context: Self::Context,
         _config: &Self::Config,
-        _target: &T,
-    ) -> impl Future<Output = bool> + Send
-    where
-        T: Target<Family = Self::Family, Digest = Self::Digest>,
-    {
+        _target: &crate::qmdb::sync::Target<Self::Family, Self::Digest>,
+    ) -> impl Future<Output = bool> + Send {
         async { false }
     }
 
-    /// Get the root of the operation log for range verification.
-    fn ops_root(&self) -> Self::Digest;
-
-    /// Get the database root digest.
+    /// Get the root digest of the database for verification
     fn root(&self) -> Self::Digest;
 }
