+++ storage/src/qmdb/sync/mod.rs
@@ -20,9 +20,6 @@ pub(crate) use database::{Config as DatabaseConfig, Database};
 pub mod resolver;
 pub(crate) use resolver::{FetchResult, Resolver};
 
-#[cfg(test)]
-pub(crate) mod tests;
-
 mod target;
 pub use target::Target;
 
@@ -45,14 +42,13 @@ where
 }
 
 /// Create/open a database and sync it to a target state
-pub async fn sync<DB, R, T>(
-    config: Config<DB, R, T>,
+pub async fn sync<DB, R>(
+    config: Config<DB, R>,
 ) -> Result<DB, Error<DB::Family, R::Error, DB::Digest>>
 where
     DB: Database,
     DB::Op: Encode,
     R: DbResolver<DB>,
-    T: Target<Family = DB::Family, Digest = DB::Digest>,
 {
     Engine::new(config).await?.sync().await
 }
